Add generate-inventory-from-netbox script
This script generates an /inventory.pre/netbox.hosts
file in which all hosts defined in the netbox are stored.
The tags assigned to the hosts are used as host groups.

This script is necessary to be able to work with the
inventory defined in cfg-generics without importing it
into Netbox even when Netbox is activated.

One issue that still needs to be fixed is updating the
inventory via the reconciler when changes are made in the
Netbox.

This will be solved later via the python-osism API/Netbox
integration. When changes are made in the Netbox, a webhook
is triggered on the API, which in turn triggers an update
of the inventory.

Signed-off-by: Christian Berendt <berendt@osism.tech>